+++ database/migrations/2025_08_04_211845_create_stock_movements_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
             $table->enum('type', ['in', 'out']);
-            $table->decimal('quantity', 8, 3);
-            $table->decimal('cost_price', 8, 2)->nullable();
+            $table->integer('quantity');
+            $table->integer('cost_price')->nullable();
             $table->text('note')->nullable();
             $table->timestamps();
         });
